requirements.txt.txt
aiogram
gigachat
python-dotenv
aiohttp